add useSingleRowStateSelection hook, remove checkbox css variables to fix outline of focus
