+++ packages/base/src/styling/CssSizeVariables.ts
@@ -42,7 +42,7 @@ export const cssVariablesStyles = `
   --${CssSizeVariablesNames.sapWcrAnalyticalTableTreePaddingLevel1}:1.5rem;
   --${CssSizeVariablesNames.sapWcrAnalyticalTableTreePaddingLevel2}:2.25rem;
   --${CssSizeVariablesNames.sapWcrAnalyticalTableTreePaddingLevel3}:2.75rem;
-  --${CssSizeVariablesNames.sapWcrAnalyticalTableSelectionColumnWidth}:48px;
+  --${CssSizeVariablesNames.sapWcrAnalyticalTableSelectionColumnWidth}:55px;
 
 }
 
@@ -63,6 +63,6 @@ export const cssVariablesStyles = `
   --${CssSizeVariablesNames.sapWcrAnalyticalTableTreePaddingLevel1}:1rem;
   --${CssSizeVariablesNames.sapWcrAnalyticalTableTreePaddingLevel2}:1.5rem;
   --${CssSizeVariablesNames.sapWcrAnalyticalTableTreePaddingLevel3}:2rem;
-  --${CssSizeVariablesNames.sapWcrAnalyticalTableSelectionColumnWidth}:36px;
+  --${CssSizeVariablesNames.sapWcrAnalyticalTableSelectionColumnWidth}:40px;
 }
 `;
