+++ packages/main/src/components/AnalyticalTable/index.tsx
@@ -236,8 +236,8 @@ const AnalyticalTable: FC<TableProps> = forwardRef((props: TableProps, ref: Ref<
     useTableHeaderGroupStyling,
     useTableHeaderStyling,
     useTableRowStyling,
-    useSingleRowStateSelection,
     useRowSelectionColumn,
+    useSingleRowStateSelection,
     useRowHighlight,
     useDynamicColumnWidths,
     useColumnsDependencies,
